Suggested refactor
