+++ src/createAnimatedComponent.tsx
@@ -7,12 +7,7 @@ import type {
   Ref,
 } from 'react';
 import React from 'react';
-import {
-  findNodeHandle,
-  NativeEventEmitter,
-  Platform,
-  StyleSheet,
-} from 'react-native';
+import { findNodeHandle, Platform, StyleSheet } from 'react-native';
 import WorkletEventHandler from './reanimated2/WorkletEventHandler';
 import setAndForwardRef from './setAndForwardRef';
 import './reanimated2/layoutReanimation/animationsManager';
@@ -58,25 +53,15 @@ import NativeReanimatedModule from './reanimated2/NativeReanimated';
 import { isSharedValue } from './reanimated2';
 import type { AnimateProps } from './reanimated2/helperTypes';
 import { removeFromPropsRegistry } from './reanimated2/PropsRegistry';
-import ReanimatedModule from './ReanimatedModule';
+import {
+  NODE_MAPPING,
+  ReanimatedEventEmitter,
+  listener,
+} from './ReanimatedModule';
 
 function dummyListener() {
   // empty listener we use to assign to listener properties for which animated
   // event is used.
-}
-
-const NODE_MAPPING = new Map();
-
-const ReanimatedEventEmitter = new NativeEventEmitter(ReanimatedModule);
-
-interface ListenerData {
-  viewTag: number;
-  props: StyleProps;
-}
-
-function listener(data: ListenerData) {
-  const component = NODE_MAPPING.get(data.viewTag);
-  component && component._updateFromNative(data.props);
 }
 
 function maybeBuild(
